Merge sequential same-mode updates
When dequeueing the next update, look ahead into the queue to see if
there are other updates with the same mode. If there are, merge them
in the current update.

This reduces the latency between update queue time and processing time,
which could have a big impact in scenarios where lots of small updates
are coming at a fast rate (e.g. when drawing on the screen).

Fixes #1.